# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mahfil Fund API",
    "version": "2.0.0",
    "description": "API-owned authentication and path-scoped multi-tenant operations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/platform/users": {
      "get": {
        "operationId": "listPlatformUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Super-admin account list"
          }
        }
      }
    },
...
  }
}